engine/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "trading_engine"
version = "0.1.0"
description = "Motor de backtest y ejecución (ejemplo minimal)"
requires-python = ">=3.10"
dependencies = [
    "pandas>=1.5",
    "numpy>=1.24",
    "loguru>=0.7.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "yfinance>=0.2.28",
    "sqlalchemy>=2.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "alpaca-py>=0.9.0",
    "websockets>=12.0",
]
